requirements.txt
Django==1.8.6
PyJWT==1.4.0
argparse==1.2.1
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.9.0
djangorestframework==3.3.1
oauthlib==1.0.1
python-openid==2.2.5
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
wsgiref==0.1.2
